backend/pyproject.toml
[project]
name = "hyacine-gallery-backend"
version = "0.1.0"
description = "Hyacine Gallery backend API"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "aiosqlite>=0.20",
    "alembic>=1.14",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "httpx[http2]>=0.28",
    "pillow>=11.1",
    "python-multipart>=0.0.18",
    "gallery-dl>=1.29",
    "imagehash>=4.3",
    "boto3>=1.42.73",
    "pyjwt>=2.9",
    "webauthn>=2.3",
    "numpy>=2.0",
]

[project.optional-dependencies]
ai = ["sentence-transformers>=3.0", "torch", "transformers>=4.40"]
dev = [
    "ruff>=0.9",
    "mypy>=1.14",
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "respx>=0.22",
]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.14"
strict = true
plugins = ["pydantic.mypy"]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"
